backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
bleach==6.2.0
celery==5.6.2
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
Django==6.0.3
django-axes==8.3.1
django-cors-headers==4.9.0
django-encrypted-model-fields==0.6.5
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
environ==1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
PyJWT==2.12.1
python-dateutil==2.9.0.post0
qrcode==8.2
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
